requirements.txt
django
requests
djangorestframework
pillow